migrate columns to statement list table
