+++ packages/client/src/pages/MainPage/containers/StatementsListBox/StatementListBox.tsx
@@ -84,7 +84,7 @@ export const StatementListBox: React.FC = () => {
     }
   );
 
-  const { statements, entities } = data || initialData;
+  const { statements, entities, right } = data || initialData;
 
   const { data: audits, isFetching: isFetchingAudits } = useQuery(
     ["territory", "statement-list", "audits", territoryId, statementListOpened],
@@ -262,207 +262,6 @@ export const StatementListBox: React.FC = () => {
     }
   };
 
-  const columns: Column<{}>[] = useMemo(() => {
-    return [
-      {
-        Header: "ID",
-        accessor: "id",
-      },
-      {
-        Header: "",
-        id: "Statement",
-        Cell: ({ row }: Cell) => {
-          const statement = row.original as IStatement;
-          return (
-            <EntityTag
-              entity={statement as IEntity}
-              showOnly="entity"
-              tooltipText={statement.data.text}
-            />
-          );
-        },
-      },
-      {
-        Header: "Subj.",
-        accessor: "data",
-        Cell: ({ row }: Cell) => {
-          const subjectIds: string[] = row.values.data?.actants
-            ? row.values.data.actants
-                .filter((a: any) => a.position === "s")
-                .map((a: any) => a.entityId)
-            : [];
-          const subjectObjects = subjectIds.map(
-            (actantId: string) => entities[actantId]
-          );
-          const definedSubjects = subjectObjects.filter((s) => s !== undefined);
-
-          return (
-            <>
-              {definedSubjects ? (
-                <TagGroup definedEntities={definedSubjects} />
-              ) : (
-                <div />
-              )}
-            </>
-          );
-        },
-      },
-      {
-        Header: "Actions",
-        Cell: ({ row }: Cell) => {
-          const actionIds = row.values.data?.actions
-            ? row.values.data.actions.map((a: any) => a.actionId)
-            : [];
-          const actionObjects: IAction[] = actionIds.map(
-            (actionId: string) => entities[actionId]
-          );
-          const definedActions = actionObjects.filter((a) => a !== undefined);
-
-          return (
-            <>
-              {definedActions ? (
-                <TagGroup definedEntities={definedActions} />
-              ) : (
-                <div />
-              )}
-            </>
-          );
-        },
-      },
-      {
-        Header: "Objects",
-        Cell: ({ row }: Cell) => {
-          const actantIds = row.values.data?.actants
-            ? row.values.data.actants
-                .filter((a: any) => a.position !== "s")
-                .map((a: any) => a.entityId)
-            : [];
-          const actantObjects: IEntity[] = actantIds.map(
-            (actantId: string) => entities[actantId]
-          );
-          const definedObjects = actantObjects.filter((o) => o !== undefined);
-
-          return (
-            <>
-              {definedObjects ? (
-                <TagGroup definedEntities={definedObjects} oversizeLimit={4} />
-              ) : (
-                <div />
-              )}
-            </>
-          );
-        },
-      },
-      {
-        Header: "Text",
-        Cell: ({ row }: Cell) => {
-          const { text } = row.values.data;
-          const maxWordsCount = 20;
-          const trimmedText = text.split(" ").slice(0, maxWordsCount).join(" ");
-          if (text?.match(/(\w+)/g)?.length > maxWordsCount) {
-            return <StyledText>{trimmedText}...</StyledText>;
-          }
-          return <StyledText>{trimmedText}</StyledText>;
-        },
-      },
-      {
-        id: "lastEdit",
-        Header: "Edited",
-        Cell: ({ row }: Cell) => {
-          return false;
-        },
-      },
-      {
-        Header: "",
-        id: "expander",
-        width: 300,
-        Cell: ({ row }: Cell) => {
-          return (
-            <ButtonGroup>
-              {data?.right !== UserEnums.RoleMode.Read && (
-                <StatementListContextMenu
-                  buttons={[
-                    <Button
-                      key="r"
-                      icon={<FaTrashAlt size={14} />}
-                      color="danger"
-                      tooltipLabel="delete"
-                      onClick={() => {
-                        setStatementToDelete(
-                          row.original as IResponseStatement
-                        );
-                        setShowSubmit(true);
-                      }}
-                    />,
-                    <Button
-                      key="d"
-                      icon={<FaClone size={14} />}
-                      color="warning"
-                      tooltipLabel="duplicate"
-                      onClick={() => {
-                        duplicateStatement(row.original as IResponseStatement);
-                      }}
-                    />,
-                    <Button
-                      key="add-up"
-                      icon={
-                        <>
-                          <FaPlus size={14} />
-                          <BsArrowUp size={14} />
-                        </>
-                      }
-                      tooltipLabel="add new statement before"
-                      color="info"
-                      onClick={() => {
-                        addStatementAtCertainIndex(row.index);
-                      }}
-                    />,
-                    <Button
-                      key="add-down"
-                      icon={
-                        <>
-                          <FaPlus size={14} />
-                          <BsArrowDown size={14} />
-                        </>
-                      }
-                      tooltipLabel="add new statement after"
-                      color="success"
-                      onClick={() => {
-                        addStatementAtCertainIndex(row.index + 1);
-                      }}
-                    />,
-                  ]}
-                />
-              )}
-              <span
-                {...row.getToggleRowExpandedProps()}
-                style={{
-                  cursor: "pointer",
-                  display: "flex",
-                  alignItems: "center",
-                }}
-                onClick={(e: React.MouseEvent) => {
-                  e.stopPropagation();
-                  const newObject = {
-                    ...rowsExpanded,
-                    [row.values.id]: !rowsExpanded[row.values.id],
-                  };
-                  dispatch(setRowsExpanded(newObject));
-                }}
-              >
-                {rowsExpanded[row.values.id] ? (
-                  <FaChevronCircleUp />
-                ) : (
-                  <FaChevronCircleDown />
-                )}
-              </span>
-            </ButtonGroup>
-          );
-        },
-      },
-    ];
-  }, [data, statementId, rowsExpanded]);
-
   const actantsUpdateMutation = useMutation(
     async (statementObject: { statementId: string; data: {} }) =>
       await api.entityUpdate(statementObject.statementId, {
@@ -514,12 +313,16 @@ export const StatementListBox: React.FC = () => {
           <StatementListTable
             statements={statements}
             audits={auditsData}
-            columns={columns}
             handleRowClick={(rowId: string) => {
               setStatementId(rowId);
             }}
             actantsUpdateMutation={actantsUpdateMutation}
             entities={entities}
+            right={right}
+            duplicateStatement={duplicateStatement}
+            setStatementToDelete={setStatementToDelete}
+            setShowSubmit={setShowSubmit}
+            addStatementAtCertainIndex={addStatementAtCertainIndex}
           />
         </StyledTableWrapper>
       ) : (
